HSZK fixes

+++ infoc-engine/src/infoc/renderer/gl.c
@@ -202,7 +202,7 @@ void opengl_shutdown()
 
 void opengl_debug(GLenum source, GLenum type, GLuint id, GLenum severity, GLsizei length, const GLchar* message, const void* userParam)
 {
-	switch (severity)
+	switch ((uint32_t)severity)
 	{
 		case GL_DEBUG_SEVERITY_NOTIFICATION:	printf("[OpenGL notification] %s\n", message); return;
 		case GL_DEBUG_SEVERITY_LOW:				printf("[OpenGL information] %s\n", message); return;
